+++ packages/hoppscotch-app/src/layouts/default.vue
@@ -57,7 +57,7 @@
 </template>
 
 <script setup lang="ts">
-import { computed, onBeforeMount, ref, watch } from "vue"
+import { computed, onBeforeMount, onMounted, ref, watch } from "vue"
 import { breakpointsTailwind, useBreakpoints } from "@vueuse/core"
 import { Splitpanes, Pane } from "splitpanes"
 import "splitpanes/dist/splitpanes.css"
@@ -66,6 +66,9 @@ import { useSetting } from "@composables/settings"
 import { defineActionHandler } from "~/helpers/actions"
 import { hookKeybindingsListener } from "~/helpers/keybindings"
 import { applySetting } from "~/newstore/settings"
+import { getLocalConfig, setLocalConfig } from "~/newstore/localpersistence"
+import { useToast } from "~/composables/toast"
+import { useI18n } from "~/composables/i18n"
 
 const router = useRouter()
 
@@ -81,10 +84,39 @@ const columnLayout = useSetting("COLUMN_LAYOUT")
 const breakpoints = useBreakpoints(breakpointsTailwind)
 const mdAndLarger = breakpoints.greater("md")
 
+const toast = useToast()
+const t = useI18n()
+
 onBeforeMount(() => {
   if (!mdAndLarger.value) {
     rightSidebar.value = false
     columnLayout.value = true
+  }
+})
+
+onMounted(() => {
+  const cookiesAllowed = getLocalConfig("cookiesAllowed") === "yes"
+  if (!cookiesAllowed) {
+    toast.show(`${t("app.we_use_cookies")}`, {
+      duration: 0,
+      action: [
+        {
+          text: `${t("action.learn_more")}`,
+          onClick: (_, toastObject) => {
+            setLocalConfig("cookiesAllowed", "yes")
+            toastObject.goAway(0)
+            window.open("https://docs.hoppscotch.io/privacy", "_blank")?.focus()
+          },
+        },
+        {
+          text: `${t("action.dismiss")}`,
+          onClick: (_, toastObject) => {
+            setLocalConfig("cookiesAllowed", "yes")
+            toastObject.goAway(0)
+          },
+        },
+      ],
+    })
   }
 })
 
@@ -94,14 +126,6 @@ watch(mdAndLarger, () => {
     rightSidebar.value = false
     columnLayout.value = true
   }
-})
-
-defineActionHandler("modals.search.toggle", () => {
-  showSearch.value = !showSearch.value
-})
-
-defineActionHandler("modals.support.toggle", () => {
-  showSupport.value = !showSupport.value
 })
 
 const spacerClass = computed(() => {
@@ -120,6 +144,14 @@ const spacerClass = computed(() => {
     return "spacer-expand"
 
   return ""
+})
+
+defineActionHandler("modals.search.toggle", () => {
+  showSearch.value = !showSearch.value
+})
+
+defineActionHandler("modals.support.toggle", () => {
+  showSupport.value = !showSupport.value
 })
 
 defineActionHandler("navigation.jump.rest", () => {
